Extract remote web pairing challenge store
